% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_check.R
\name{delete_factor}
\alias{delete_factor}
\title{Delete the kth dimension from res object}
\usage{
delete_factor(res, k, Y, a, b, S2, var.type, S2.type)
}
\description{
Delete the kth dimension from res object
}
