% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funflash.R
\name{funflash}
\alias{funflash}
\title{Fit smoothing flash}
\usage{
funflash(
  data,
  var.type = "by_row",
  Kmax = 10,
  tol = 0.01,
  maxiter = 1000,
  ebnm_fn = "ebnm_pn",
  init_fn = "udv_si",
  ebnm_param = NULL,
  verbose = TRUE,
  nullcheck = TRUE,
  sigma2 = NULL,
  seed = 12345
)
}
\arguments{
\item{data}{a data obj from funflash_set_data()}

\item{var.type}{"by_column","by_row", "constant", "zero".}

\item{Kmax}{max number of factors}

\item{tol}{tol for stopping the iterations}

\item{maxiter}{max number of iteration}

\item{ebnm_fn}{'ebnm_pn', 'ebnm_pl', 'ebnm_ash'}

\item{init_fn}{'udv_si',..., see ?flashr}

\item{nullcheck}{whether perform null check}

\item{sigma2}{if not null and var.type='zero', treat it as given. Can be used as test purpose.}
}
\value{
a list of
 \item{ldf}{a list of d, El, Ef}
 \item{Fb}{E(Fb) before dividing b}
 \item{nfactors}{number of factors}
 \item{pve}{proportion/percentage variance explained}
 \item{fitted.values}{aELEF'b}
 \item{fit}{result list, before transforming back to data space}
 \item{objective}{objective funtion value}
}
\description{
Fit smoothing flash
}
