% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funflash_set_data.R
\name{funflash_set_data}
\alias{funflash_set_data}
\title{Set data for funflash}
\usage{
funflash_set_data(
  Y,
  a = NULL,
  b = NULL,
  S2 = NULL,
  family = "DaubExPhase",
  filter.number = 1,
  type = "wavelet",
  reflect.data = FALSE
)
}
\arguments{
\item{Y}{data matrix}

\item{a}{row scaling factors}

\item{b}{column scaling factors, should be smooth}

\item{S2}{known variance. Can be a scalar, a vector or a matrix}

\item{family, filter.number}{wavelet family and filter number as in wavethresh package}

\item{type}{type of wavelet transformation, "wavelet" or "station". Choose 'station' could give better fit of the smooth factor but would introduce redundant factors.}

\item{reflect.data}{whether reflect the data matrix. Setting this to TRUE would give better fit to smooth F at boundary, but would likely introduce redundant factors.}
}
\value{
a list of:
 \item{Y}{wavelet coefficients matrix, excluding the energy term}
 \item{a}{row scaling factor}
 \item{b}{a 1 vector. Because we need to DWT the bF together. Treat bF as the F and will divide b to get original F}
 \item{S2,S2.type}{S2 the known variance matrix, and its type}
 \item{orig.idx}{original index of Y before making it a power of 2 and possible reflection}
 \item{nlevel}{number of levels}
 \item{energy}{scaled summation of each row of Y}
 \item{orig.p}{number of column of Y after reflect() before DWT}
 \item{orig.b}{the input b}
 \item{orig.Y}{the original Y matrix(after reflection/power of 2)}
 \item{energy.S2}{the S2 for energy vector}
}
\description{
Set data for funflash
}
