% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funflash_set_data.R
\name{get_wave_index}
\alias{get_wave_index}
\title{Obtain index(position of coefficients in a vector) of each scale.}
\usage{
get_wave_index(t, nt, type)
}
\arguments{
\item{t}{current wavelet level, take value 0:(nt-1), 0 is the coasest level}

\item{nt}{total number of level}

\item{type}{'wavelet', 'station'}
}
\description{
Obtain index(position of coefficients in a vector) of each scale.
}
