% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_check.R
\name{is_tiny_fl}
\alias{is_tiny_fl}
\title{Check if the added loading and factors are very small. If so, break the iteration.}
\usage{
is_tiny_fl(f, k, tol = 1e-08)
}
\description{
Check if the added loading and factors are very small. If so, break the iteration.
}
