% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledflash.R
\name{scaledflash}
\alias{scaledflash}
\title{Fit scaled flash}
\usage{
scaledflash(
  Y,
  a = NULL,
  b = NULL,
  S2 = NULL,
  var.type = "by_row",
  Kmax = 50,
  tol = 0.01,
  maxiter = 1000,
  ebnm_fn = "ebnm_pn",
  init_fn = "udv_si",
  ebnm_param = NULL,
  verbose = TRUE,
  nullcheck = TRUE,
  sigma2 = NULL,
  seed = 12345
)
}
\arguments{
\item{Y}{a data obj from funflash_set_data()}

\item{a}{row scaling factors}

\item{b}{column scaling factors}

\item{S2}{Known variances. Can be NULL, a scalar, a vector or a matrix.}

\item{var.type}{whether assume F is smooth.}
}
\description{
Fit scaled flash
}
