% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledflash_rank1.R
\name{scaledflash_rank1}
\alias{scaledflash_rank1}
\title{Fit rank 1 scaled flash}
\usage{
scaledflash_rank1(
  Rk,
  a,
  b,
  S2,
  res,
  k,
  var.type,
  ebnm_fn,
  ebnm_param,
  maxiter = 100,
  tol = 0.01,
  verbose,
  init_fn,
  Y,
  S2.type
)
}
\description{
Fit rank 1 scaled flash
}
