% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biwhitening.R
\name{Sinkhorn_Knopp}
\alias{Sinkhorn_Knopp}
\title{Sinkhorn_Knopp for matrix row and col scaling}
\usage{
Sinkhorn_Knopp(A, rs = NULL, cs = NULL, tol = 1e-05, maxiter = 100)
}
\arguments{
\item{A}{nonnegative matrix}

\item{rs}{target row sums, a vector of length m}

\item{cs}{target col sums, a vector of length n}

\item{tol}{tol to stop the iterations}
}
\description{
Sinkhorn_Knopp for matrix row and col scaling
}
