% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_flash.R
\name{smooth_flash}
\alias{smooth_flash}
\title{Fit smoothed flash}
\usage{
smooth_flash(
  Y,
  S = NULL,
  Kmax = 10,
  tol = 0.01,
  init_fn = "udv_si",
  ebnm_fn = "ebnm_pn",
  var_type = "by_row",
  ebnm_param = NULL,
  f_init = NULL,
  filter.number = 1,
  family = "DaubExPhase",
  nullcheck = TRUE,
  maxiter = 100,
  verbose = TRUE,
  type = "wavelet",
  seed = 12345
)
}
\arguments{
\item{Y}{data matrix, n by p, p is a power of 2.}

\item{S}{An n by p matrix of the standard errors of the observations in Y. (Can be a scalar if all standard errors are equal.) If S = NULL, then the standard errors will be estimated during fitting.}

\item{Kmax}{maximum number of topics}

\item{tol}{tolerance}

\item{init_fn}{initialization of}

\item{type}{wavelet or station. Use wavelet for exact ELBO calc.}

\item{var.type}{Use 'by_row' for exact ELBO calc}
}
\value{
same as flashr
}
\description{
This model fits wavelet prior for each factors.
}
