% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sigma2.R
\name{update_sigma2}
\alias{update_sigma2}
\title{Update the variance}
\usage{
update_sigma2(Y, a, b, S2, res, var.type, S2.type, energy, energy.S2)
}
\arguments{
\item{Y}{data matrix}

\item{a, b}{scaling factors}
}
\value{
updated sigma2, a matrix
}
\description{
Update the variance
}
